[IR] Record owning graph for input/output and initializers
